Adjust reference waveform time using `zeroTime` only
Using `tvec` to adjust the reference waveform time vector means that the
timing is being modified by both `zeroTime` and `T0`, since `tvec` is
already adjusted for both of those. However, we only want to adjust the
reference waveform with `zeroTime`. Essentially, it seems like `tvec`
should only be used for plotting force time series related things (since
the force time series needs to be corrected for `T0`, but the reference
waveforms do not)